Refactor to use the command pattern
