+++ crates/cef-ui-tools/Cargo.toml
@@ -9,7 +9,7 @@ anyhow = { workspace = true }
 tracing = { workspace = true }
 tracing-log = { workspace = true }
 tracing-subscriber = { workspace = true }
-
+clap = { version = "4.0", features = ["derive"] }
 
 [[bin]]
 name = "cef-artifacts"
