+++ crates/cef-ui-util/src/commands/mod.rs
@@ -1,0 +1,9 @@
+mod app_bundle;
+mod artifacts;
+mod build;
+mod clean;
+
+pub use app_bundle::*;
+pub use artifacts::*;
+pub use build::*;
+pub use clean::*;
