+++ crates/cef-ui-util/src/lib.rs
@@ -1,11 +1,11 @@
+mod commands;
 mod download;
 mod link;
 mod paths;
-mod tools;
 mod util;
 
+pub use commands::*;
 pub use download::*;
 pub use link::*;
 pub use paths::*;
-pub use tools::*;
 pub use util::*;
